backend/Django_RESTFramework/pyproject.toml
[project]
name = "Django_RESTFramework"
version = "0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Foundation_Building"
version = "0.1.0"
packages = [
  { include = "Foundation_Building/Foundation_Building" },
  { include = "Foundation_Building/library" },
]
authors = ["kamegin4-aws <kamegin4@outlook.jp>"]
description = "Backend App of Foundation_Building"
readme = "README.md"
license = "MIT"
#keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
]

[tool.poetry.dependencies]
python = "^3.12.1"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pygments = "^2.18.0"
django-filter = "^24.2"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
